Move flexbox classes into a flexbox subpackage.
