+++ redwood-flexbox-engine/src/commonMain/kotlin/app/cash/redwood/flexbox/Direction.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood
+package app.cash.redwood.flexbox
 
 internal fun FlexDirection.toDirection(): Direction {
   if (this == FlexDirection.Row || this == FlexDirection.RowReverse) {
