+++ redwood-flexbox-engine/src/commonMain/kotlin/app/cash/redwood/flexbox/Node.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood
+package app.cash.redwood.flexbox
 
 /**
  * A node with properties that can be measured and laid out inside a flexbox.
