+++ redwood-flexbox-engine/src/commonMain/kotlin/app/cash/redwood/flexbox/properties.kt
@@ -15,7 +15,7 @@
  */
 @file:Suppress("MemberVisibilityCanBePrivate")
 
-package app.cash.redwood
+package app.cash.redwood.flexbox
 
 import kotlin.jvm.JvmInline
 import kotlin.math.abs
