+++ redwood-flexbox-engine/src/commonTest/kotlin/app/cash/redwood/flexbox/FlexboxEngineTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Google Inc. All rights reserved.
+ * Copyright (C) 2022 Square, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood
+package app.cash.redwood.flexbox
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
