+++ redwood-flexbox-engine/src/commonTest/kotlin/app/cash/redwood/flexbox/FlexboxTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood
+package app.cash.redwood.flexbox
 
 import app.cash.redwood.FlexDirection.Companion.Row
 import app.cash.redwood.FlexDirection.Companion.RowReverse
